# specs/billing/openapi.yaml
openapi: 3.1.0
info:
  title: Billing API
  version: 1.0.0
  description: |
    Billing module for invoice lifecycle management.
    Supports tenant-scoped invoice creation, retrieval, updates, and voiding.
  contact:
    name: Pcampus Studio
    url: https://github.com/pcampus-studio/pcampus-openapi-kit

servers:
  - url: https://api.example.com/v1
    description: Production billing API

tags:
  - name: Invoices
    description: Manage billing invoices for authenticated tenants.

paths:
  /invoices:
    get:
      tags: [Invoices]
      operationId: list_invoices
      summary: List invoices
      description: |
        Retrieve a cursor-paginated list of invoices for the authenticated tenant.
        Returns 401 when authentication is missing or invalid.
        Returns 403 when the requester lacks tenant access.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '../shared/pagination.yaml#/components/parameters/Cursor'
        - $ref: '../shared/pagination.yaml#/components/parameters/Limit'
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [draft, open, paid, void]
          description: Filter invoices by lifecycle status.
      responses:
        '200':
          description: Paginated list of invoices.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedInvoiceList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      x-agent-hints:
        safe_to_retry: true
        destructive: false
        requires_confirmation: false
        workflow_hint: Use this endpoint to discover invoices before fetching details or applying updates.

    post:
      tags: [Invoices]
      operationId: create_invoice
      summary: Create invoice
      description: |
        Create a draft invoice for the authenticated tenant.
        Returns 422 when line items or customer reference fail validation.
        Returns 409 when invoice_number already exists in the tenant.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceCreateRequest'
      responses:
        '201':
          description: Invoice created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
      x-agent-hints:
        safe_to_retry: false
        destructive: false
        requires_confirmation: true
        workflow_hint: Confirm customer_id and line items with the requester before creating an invoice.

  /invoices/{invoice_id}:
    get:
      tags: [Invoices]
      operationId: get_invoice
      summary: Get invoice
      description: |
        Retrieve a single invoice by UUID for the authenticated tenant.
        Returns 404 when the invoice does not exist or is not visible in the tenant.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/InvoiceId'
      responses:
        '200':
          description: Invoice found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      x-agent-hints:
        safe_to_retry: true
        destructive: false
        requires_confirmation: false

    patch:
      tags: [Invoices]
      operationId: update_invoice
      summary: Update invoice
      description: |
        Partially update a draft or open invoice by UUID.
        Returns 404 when the invoice does not exist.
        Returns 409 when attempting to modify a void or paid invoice.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/InvoiceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceUpdateRequest'
      responses:
        '200':
          description: Invoice updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
      x-agent-hints:
        safe_to_retry: false
        destructive: false
        requires_confirmation: true

    delete:
      tags: [Invoices]
      operationId: void_invoice
      summary: Void invoice
      description: |
        Void an invoice by UUID. This is a destructive lifecycle action.
        Returns 404 when the invoice does not exist.
        Returns 409 when the invoice is already paid or void.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/InvoiceId'
      responses:
        '204':
          description: Invoice voided successfully.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
      x-agent-hints:
        safe_to_retry: false
        destructive: true
        requires_confirmation: true
        workflow_hint: Explain that voiding prevents further collection on the invoice.

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token issued by the platform identity service.

  parameters:
    InvoiceId:
      name: invoice_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier of the invoice.

  schemas:
    InvoiceStatus:
      type: string
      enum:
        - draft
        - open
        - paid
        - void
      description: Current lifecycle status of the invoice.

    InvoiceLineItem:
      type: object
      required:
        - description
        - quantity
        - unit_amount
      properties:
        description:
          type: string
        quantity:
          type: integer
          minimum: 1
        unit_amount:
          type: number
          format: double
          minimum: 0

    Invoice:
      type: object
      required:
        - id
        - customer_id
        - invoice_number
        - status
        - currency
        - amount_due
        - line_items
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        invoice_number:
          type: string
          example: INV-2026-0001
        status:
          $ref: '#/components/schemas/InvoiceStatus'
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: THB
        amount_due:
          type: number
          format: double
          minimum: 0
        due_date:
          type: string
          format: date
          nullable: true
        line_items:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceLineItem'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    InvoiceCreateRequest:
      type: object
      required:
        - customer_id
        - invoice_number
        - currency
        - line_items
      properties:
        customer_id:
          type: string
          format: uuid
        invoice_number:
          type: string
        currency:
          type: string
          minLength: 3
          maxLength: 3
        due_date:
          type: string
          format: date
          nullable: true
        line_items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/InvoiceLineItem'

    InvoiceUpdateRequest:
      type: object
      properties:
        due_date:
          type: string
          format: date
          nullable: true
        line_items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/InvoiceLineItem'
        status:
          type: string
          enum:
            - open
          description: Only `open` is allowed when transitioning from draft.

    PaginatedInvoiceList:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Invoice'
        pagination:
          $ref: '../shared/pagination.yaml#/components/schemas/PaginationMeta'

  responses:
    Unauthorized:
      description: Authentication is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '../shared/errors.yaml#/components/schemas/ErrorResponse'
    Forbidden:
      description: Requester lacks access to the invoice or tenant.
      content:
        application/json:
          schema:
            $ref: '../shared/errors.yaml#/components/schemas/ErrorResponse'
    NotFound:
      description: Invoice does not exist.
      content:
        application/json:
          schema:
            $ref: '../shared/errors.yaml#/components/schemas/ErrorResponse'
    Conflict:
      description: Request conflicts with current invoice state.
      content:
        application/json:
          schema:
            $ref: '../shared/errors.yaml#/components/schemas/ErrorResponse'
    ValidationError:
      description: Input failed validation.
      content:
        application/json:
          schema:
            $ref: '../shared/errors.yaml#/components/schemas/ValidationErrorResponse'

security:
  - bearerAuth: []